Fix imports, merge with latest name changes of pool-system
